add: Added events for total bot and members vc
